+++ lib/activity.ml
@@ -21,6 +21,73 @@ let title { title; _ } = title
 
 type t = { projects : project list; activity : Get_activity.Contributions.t }
 
+(** Grouping the activity items together in a tree: 1 item can have many sub-items *)
+module Tree = struct
+  open Get_activity.Contributions
+
+  type t = (item * item list) list Repo_map.t
+
+  let order_item_kind = function
+    | `New_repo -> 0
+    | `Issue -> 1
+    | `PR -> 2
+    | `Merge -> 3
+    | `Review _ -> 4
+    | `Comment `PR -> 5
+    | `Comment `Issue -> 6
+
+  let sort_items x y =
+    match Int.compare (order_item_kind x.kind) (order_item_kind y.kind) with
+    | 0 -> String.compare x.date y.date
+    | cmp -> cmp
+
+  let root_of_url url =
+    match String.split_on_char '#' url with
+    | [ root_url ] -> root_url
+    | [ root_url; _comment_suffix ] -> root_url
+    | _ ->
+        Logs.warn (fun m -> m "invalid url: %s" url);
+        url
+
+  let contains match_kind url (c, _) =
+    match_kind c.kind && String.equal (root_of_url url) (root_of_url c.url)
+
+  let add_to_maybe_another_item ~can_add_to item acc =
+    let found, rest = List.partition (contains can_add_to item.url) acc in
+    match found with
+    (* the comment is stored as a root item *)
+    | [] -> (item, []) :: rest
+    (* the comment is stored as a subitem of another item *)
+    | (root, subitems) :: _ -> (root, item :: subitems) :: rest
+
+  let make activity : t =
+    activity
+    |> Repo_map.map (fun items ->
+           List.sort sort_items items
+           |> List.fold_left
+                (fun acc item ->
+                  match item.kind with
+                  | `Issue | `PR | `New_repo | `Review _ -> (item, []) :: acc
+                  | `Merge ->
+                      add_to_maybe_another_item item acc ~can_add_to:(function
+                        | `PR -> true
+                        | _ -> false)
+                  | `Comment `PR ->
+                      add_to_maybe_another_item item acc ~can_add_to:(function
+                        | `PR | `Review _ | `Merge | `Comment `PR -> true
+                        | _ -> false)
+                  | `Comment `Issue ->
+                      add_to_maybe_another_item item acc ~can_add_to:(function
+                        | `Issue | `Comment `Issue -> true
+                        | _ -> false))
+                []
+           |> List.rev)
+end
+
+let pp_kind ppf = function
+  | `Issue -> Fmt.string ppf "issue"
+  | `PR -> Fmt.string ppf "PR"
+
 let pp_last_week username ppf projects =
   let pp_items ppf t =
     let t = if t = [] then [ "Work Item 1" ] else t in
@@ -48,34 +115,42 @@ let repo_org ?(with_id = false) ?(no_links = false) f s =
   | repo :: org :: _ -> Fmt.pf f "[%s/%s](%s)" org repo s
   | _ -> Fmt.failwith "Malformed URL %S" s
 
-let pp_ga_item ?(gitlab = false) ~no_links () f
+let pp_ga_item ?(gitlab = false) ~no_links sub_items f
     (t : Get_activity.Contributions.item) =
   match gitlab with
   | true -> (
       match t.kind with
-      | `PR -> Fmt.pf f "PR (Gitlab): %s %s" t.title t.url
-      | `Issue -> Fmt.pf f "Issue (Gitlab): %s %s" t.title t.url
+      | `PR -> Fmt.pf f "Opened PR (Gitlab): %s %s" t.title t.url
+      | `Issue -> Fmt.pf f "Opened issue (Gitlab): %s %s" t.title t.url
       | _ -> ())
   | false -> (
       match t.kind with
       | `Issue ->
-          Fmt.pf f "Issue: %s %a" t.title
+          Fmt.pf f "Opened issue: %s %a" t.title
             (repo_org ~with_id:true ~no_links)
             t.url
       | `PR ->
-          Fmt.pf f "PR: %s %a" t.title (repo_org ~with_id:true ~no_links) t.url
-      | `Comment `Issue ->
-          Fmt.pf f "Commented on issue %S %a" t.title
+          let merged =
+            List.exists
+              (fun x -> x.Get_activity.Contributions.kind = `Merge)
+              sub_items
+          in
+          let status = if merged then "Opened and merged" else "Opened" in
+          Fmt.pf f "%s PR: %s %a" status t.title
             (repo_org ~with_id:true ~no_links)
             t.url
-      | `Comment `PR ->
-          Fmt.pf f "Commented on PR %S %a" t.title
+      | `Comment kind ->
+          Fmt.pf f "Commented on %a: %s %a" pp_kind kind t.title
             (repo_org ~with_id:true ~no_links)
             t.url
       | `Review s ->
-          Fmt.pf f "%s %s %a" s t.title (repo_org ~with_id:true ~no_links) t.url
+          Fmt.pf f "%s PR: %s %a"
+            (String.capitalize_ascii @@ String.lowercase_ascii s)
+            t.title
+            (repo_org ~with_id:true ~no_links)
+            t.url
       | `Merge ->
-          Fmt.pf f "Merged %S %a" t.title
+          Fmt.pf f "Merged PR: %s %a" t.title
             (repo_org ~with_id:true ~no_links)
             t.url
       | `New_repo ->
@@ -85,8 +160,9 @@ let pp_ga_item ?(gitlab = false) ~no_links () f
 
 let pp_activity ?(gitlab = false) ~no_links () ppf activity =
   let open Get_activity.Contributions in
-  let pp_item ppf item =
-    Fmt.pf ppf "  - %a" (pp_ga_item ~gitlab ~no_links ()) item
+  let activity = Tree.make activity in
+  let pp_item ppf (item, sub_items) =
+    Fmt.pf ppf "  - %a" (pp_ga_item ~gitlab ~no_links sub_items) item
   in
   let bindings = Repo_map.bindings activity in
   let pp_binding ppf (_repo, items) =